fixes meeting detail view
